Add working controller input code
For now, the controller input logic is just in the main activity of the
app and it just works right away without waiting for login or car
selection or anything.

It sends a UDP MOVE message every tenth of a second, regardless of whether
the thumsbtick on the controller has moved since the last message. This is
to prevent any issues with dropped UDP packets.